Genotypes in tandem repeats could be large
We used np.int8 for storing family genotypes. Since we moved
to newer version of NumPy this causes an exception for some
tandem repeat variants.

We switched family genotypes to np.int64.

For Schema1 we still insist that genotypes can be stored in
signed np.int8 because of the serialization used.

+++ dae/dae/variants_loaders/vcf/loader.py
@@ -124,7 +124,7 @@ class VcfFamiliesGenotypes(FamiliesGenotypes):
             if len(family_genotype) == 0:
                 continue
 
-            genotype = np.array(family_genotype, np.int8)
+            genotype = np.array(family_genotype, np.int64)
             genotype = genotype.T
             assert len(genotype.shape) == 2, (genotype, family)
             assert genotype.shape[0] == 2
@@ -569,7 +569,7 @@ class SingleVcfLoader(VariantsGenotypesLoader):
                 assert known_independent_genotypes is not None
 
                 independent_genotypes = np.array(
-                    known_independent_genotypes, np.int8).T
+                    known_independent_genotypes, np.int64).T
 
                 self._calc_allele_frequencies(
                     current_summary_variant,
